Item 4 (medium): widen model-blob length path beyond 64 KB
encrypted_len and the BLAKE2s/ChaCha20 length parameters were uint16_t, so
any model whose encrypted blob exceeded 64 KB was silently truncated: the
length wrapped, the MAC was computed over the wrong byte count, and a valid
model was rejected (or, for a single >64 KB layer, decryption stopped
early). This contradicted the advertised STM32F4 ~800K-param target.

Widened the whole blob-length path to uint32_t: mnv_model_t.encrypted_len,
mnv_crypto_header_t.weight_count/bias_count, the BLAKE2s update/mac/verify
lengths (and internal fill handling), the ChaCha20 decrypt length and loop
index, the mnv_init PROGMEM chunk-read counters, and the per-layer
ct_offset / weight-byte counts in the MLP, CNN1D, and BNN forward passes.
Flat-memory targets (STM32, host) now handle large models correctly.

AVR near-pointer reality: pgm_read_byte uses a 16-bit flash pointer, so on
PROGMEM targets mnv_init() now rejects a blob > 64 KB with MNV_ERR_CONFIG
(before any read) rather than wrapping the address. Documented in the
Supported Targets table (ATmega2560 † footnote); >64 KB on AVR needs
far-pointer support, which is future work.

Also fixed a latent config-precedence bug surfaced while testing this: the
unconditional `#define MNV_TARGET_ATMEGA328P` meant -DMNV_TARGET_HOST was
ignored and host tests secretly ran under ATmega's 14 KB budget and PROGMEM
path. Guarded the default so an explicit -D target wins (ATmega328P still
the default when none is given).

Tests: mlp_big config runs a ~69 KB MLP through the full pipeline against an
independent reference (verified it FAILS with uint16 encrypted_len, PASSES
with uint32); test_progmem_guard covers the PROGMEM flash-read path — which
the now-true-host configs no longer exercise — and the >64 KB AVR guard.
Full suite green under ASan/UBSan.

+++ include/mnv_types.h
@@ -181,8 +181,8 @@ typedef struct {
     uint8_t  iv[MNV_CHACHA20_IV_SIZE];      /* ChaCha20 nonce             */
     uint8_t  mac[MNV_BLAKE2S_DIGEST_SIZE];  /* BLAKE2s over ciphertext    */
                                             /* (encrypt-then-MAC)         */
-    uint16_t weight_count;                  /* number of weights         */
-    uint16_t bias_count;                    /* number of biases          */
+    uint32_t weight_count;                  /* number of weights         */
+    uint32_t bias_count;                    /* number of biases          */
 } mnv_crypto_header_t;
 
 /* =========================================================================
@@ -198,7 +198,9 @@ typedef struct {
     const mnv_crypto_header_t *crypto;      /* crypto header for weight blob     */
     const uint8_t       *key;               /* 256-bit ChaCha20 key (from fuse/KDF)*/
     const uint8_t       *encrypted_weights; /* entire encrypted weight blob      */
-    uint16_t             encrypted_len;
+    uint32_t             encrypted_len;     /* blob length in bytes (up to 4 GB; */
+                                            /* AVR near-PROGMEM caps this at 64 KB,*/
+                                            /* enforced at mnv_init())            */
 } mnv_model_t;
 
 #define MNV_ABI_VERSION  0x01U
